Implemented full functionality of the user profile view
Implemented full functionality of the user profile view, implemented some fluff statistics (may need to double check and make sure that they update accordingly). Changed instances of Not-ify (group name) to Bop Swap (Product Name), and replaced the "log out" button with the "profile" button and moved the "log out" button to be in the profile section.